add fee

+++ contracts/interfaces/IWETH.sol
@@ -3,5 +3,6 @@ pragma solidity ^0.8.17;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
+
+    function transfer(address to, uint256 value) external returns (bool);
 }
